Move file to ~/lib/server
