+++ core/actions/_lib/runActionInstance.ts
@@ -13,10 +13,10 @@ import type { ClientException, ClientExceptionOptions } from "~/lib/serverAction
 import { db } from "~/kysely/database";
 import { getPub, getPubCached } from "~/lib/server";
 import { autoRevalidate } from "~/lib/server/cache/autoRevalidate";
+import { validatePubValues } from "../../lib/server/validateFields";
 import { getActionByName } from "../api";
 import { getActionRunByName } from "./getRuns";
 import { resolveWithPubfields } from "./resolvePubfields";
-import { validatePubValues } from "./validateFields";
 
 export type ActionInstanceRunResult = ClientException | ClientExceptionOptions | ActionSuccess;
