+++ core/app/components/pubs/PubEditor/actions.ts
@@ -6,7 +6,6 @@ import type { CommunitiesId, PubsId, PubTypesId, StagesId } from "db/public";
 import { logger } from "logger";
 
 import type { PubValues } from "~/lib/server";
-import { validatePubValuesBySchemaName } from "~/actions/_lib/validateFields";
 import { db } from "~/kysely/database";
 import { getLoginData } from "~/lib/auth/loginData";
 import { isCommunityAdmin } from "~/lib/auth/roles";
@@ -14,6 +13,7 @@ import { createPubRecursiveNew } from "~/lib/server";
 import { autoCache } from "~/lib/server/cache/autoCache";
 import { autoRevalidate } from "~/lib/server/cache/autoRevalidate";
 import { defineServerAction } from "~/lib/server/defineServerAction";
+import { validatePubValuesBySchemaName } from "~/lib/server/validateFields";
 
 export const createPubRecursive = defineServerAction(async function createPubRecursive(
 	...[props]: Parameters<typeof createPubRecursiveNew>
